[Digitizing] Finish logic refactoring
