+++ assets/src/components/Digitizing.js
@@ -13,6 +13,9 @@ export default class Digitizing extends HTMLElement {
         const mainTemplate = () => html`
         <div class="digitizing">
             <div class="selectiontool-query-buttons btn-group">
+                <button type="button" class="selectiontool-toggle-visibilty btn btn-mini" @click=${() => mainLizmap.digitizing.toggleFeatureDrawnVisibility()}  data-original-title="${lizDict['tree.button.checkbox']}">
+                    <i class="icon-eye-${mainLizmap.digitizing._featureDrawnVisibility ? 'open' : 'close'}"></i>
+                </button>
                 <button type="button" class="selectiontool-query-deactivate btn btn-small ${mainLizmap.digitizing.toolSelected === 'deactivate' ? 'active' : ''}" @click=${() => mainLizmap.digitizing.toolSelected = 'deactivate'} data-original-title="${lizDict['selectiontool.toolbar.query.deactivate']}">
                     <i class="icon-none qgis_sprite mIconDeselected"></i>
                 </button>
