+++ assets/src/modules/SelectionTool.js
@@ -1,7 +1,4 @@
 import {mainLizmap, mainEventDispatcher} from '../modules/Globals.js';
-import GeoJSONReader from 'jsts/org/locationtech/jts/io/GeoJSONReader.js';
-import GeoJSONWriter from 'jsts/org/locationtech/jts/io/GeoJSONWriter.js';
-import BufferOp from 'jsts/org/locationtech/jts/operation/buffer/BufferOp.js';
 
 export default class SelectionTool {
 
@@ -14,8 +11,6 @@ export default class SelectionTool {
 
         this._newAddRemove = ['new', 'add', 'remove'];
         this._newAddRemoveSelected = this._newAddRemove[0];
-
-        this._toogleSelectionLayerVisibility = true;
 
         // Verifying WFS layers
         const featureTypes = mainLizmap.vectorLayerFeatureTypes;
@@ -67,237 +62,15 @@ export default class SelectionTool {
             format => !['GML2', 'GML3', 'GEOJSON'].includes(format.tagName)
         );
 
-        // Draw and selection tools style
-        const drawStyle = new OpenLayers.Style({
-            pointRadius: 7,
-            fillColor: '#94EF05',
-            fillOpacity: 0.2,
-            strokeColor: 'yellow',
-            strokeOpacity: 1,
-            strokeWidth: 2
-        });
-
-        const drawStyleTemp = new OpenLayers.Style({
-            pointRadius: 7,
-            fillColor: 'orange',
-            fillOpacity: 0.3,
-            strokeColor: 'blue',
-            strokeOpacity: 1,
-            strokeWidth: 3
-        });
-
-        const drawStyleSelect = new OpenLayers.Style({
-            pointRadius: 7,
-            fillColor: 'blue',
-            fillOpacity: 0.3,
-            strokeColor: 'blue',
-            strokeOpacity: 1,
-            strokeWidth: 3
-        });
-
-        const drawStyleMap = new OpenLayers.StyleMap({
-            'default': drawStyle,
-            'temporary': drawStyleTemp,
-            'select': drawStyleSelect
-        });
-
-        const queryLayer = new OpenLayers.Layer.Vector(
-            'selectionQueryLayer', {
-                styleMap: drawStyleMap
-            }
-        );
-
-        const bufferLayer = new OpenLayers.Layer.Vector(
-            'selectionBufferLayer', {
-                styleMap: drawStyleMap
-            }
-        );
-
-        mainLizmap.lizmap3.map.addLayers([queryLayer, bufferLayer]);
-        mainLizmap.lizmap3.layers['selectionQueryLayer'] = queryLayer;
-        mainLizmap.lizmap3.layers['selectionBufferLayer'] = bufferLayer;
-
-        mainLizmap.lizmap3.controls['selectiontool'] = {};
-
-        const onQueryFeatureAdded = (feature) => {
-            /**
-             * @todo Ne gère que si il ya a seulement 1 géométrie
-             */
-            if (feature.layer) {
-                if (feature.layer.features.length > 1) {
-                    feature.layer.destroyFeatures(feature.layer.features.shift());
-                }
-            }
-
-            let featureToRequest = feature;
-
-            // Handle buffer if any
-            mainLizmap.lizmap3.layers['selectionBufferLayer'].destroyFeatures();
-            if (this._bufferValue > 0) {
-                const geoJSONParser = new OpenLayers.Format.GeoJSON();
-                const selectionGeoJSON = geoJSONParser.write(feature.geometry);
-                const jstsGeoJSONReader = new GeoJSONReader();
-                const jstsGeoJSONWriter = new GeoJSONWriter();
-                const jstsGeom = jstsGeoJSONReader.read(selectionGeoJSON);
-                const jstsbBufferedGeom = BufferOp.bufferOp(jstsGeom, this._bufferValue);
-                const bufferedSelection = geoJSONParser.read(jstsGeoJSONWriter.write(jstsbBufferedGeom));
-
-                // Draw buffer
-                mainLizmap.lizmap3.layers['selectionBufferLayer'].addFeatures(bufferedSelection);
-                mainLizmap.lizmap3.layers['selectionBufferLayer'].redraw(true);
-
-                featureToRequest = bufferedSelection[0];
-            }
-
-            for (const featureType of this.allFeatureTypeSelected) {
-                mainLizmap.lizmap3.selectLayerFeaturesFromSelectionFeature(featureType, featureToRequest, this._geomOperator);
-            }
-
-            this.toolSelected = 'deactivate';
-        };
-
-        /**
-         * Box
-         * @type @new;OpenLayers.Control.DrawFeature
-         */
-        const queryBoxLayerCtrl = new OpenLayers.Control.DrawFeature(queryLayer,
-            OpenLayers.Handler.RegularPolygon,
-            {handlerOptions: {sides: 4, irregular: true}, 'featureAdded': onQueryFeatureAdded}
-        );
-
-        /**
-         * Circle
-         * @type @new;OpenLayers.Control.DrawFeature
-         */
-        const queryCircleLayerCtrl = new OpenLayers.Control.DrawFeature(queryLayer,
-            OpenLayers.Handler.RegularPolygon,
-            {handlerOptions: {sides: 40}, 'featureAdded': onQueryFeatureAdded}
-        );
-
-        /**
-         * Point
-         * @type @new;OpenLayers.Control.DrawFeature
-         */
-        const queryPointLayerCtrl = new OpenLayers.Control.DrawFeature(
-            queryLayer,
-            OpenLayers.Handler.Point,
-            {
-                'featureAdded': onQueryFeatureAdded,
-                styleMap: drawStyleMap,
-                eventListeners: {
-                    // getFeatureInfo and point draw controls are mutually exclusive
-                    'activate': function() {
-                        if ('featureInfo' in lizMap.controls && lizMap.controls.featureInfo && lizMap.controls.featureInfo.active) {
-                            lizMap.controls.featureInfo.deactivate();
-                        }
-                    },
-                    'deactivate': function() {
-                        if ('featureInfo' in lizMap.controls && lizMap.controls.featureInfo && !lizMap.controls.featureInfo.active) {
-                            lizMap.controls.featureInfo.activate();
-                        }
-                    }
-                }
-            }
-        );
-
-        /**
-         * Line
-         * @type @new;OpenLayers.Control.DrawFeature
-         */
-        const queryLineLayerCtrl = new OpenLayers.Control.DrawFeature(
-            queryLayer,
-            OpenLayers.Handler.Path,
-            {
-                'featureAdded': onQueryFeatureAdded,
-                styleMap: drawStyleMap,
-                eventListeners: {
-                    // getFeatureInfo and point draw controls are mutually exclusive
-                    'activate': function() {
-                        if ('featureInfo' in lizMap.controls && lizMap.controls.featureInfo && lizMap.controls.featureInfo.active) {
-                            lizMap.controls.featureInfo.deactivate();
-                        }
-                    },
-                    'deactivate': function() {
-                        if ('featureInfo' in lizMap.controls && lizMap.controls.featureInfo && !lizMap.controls.featureInfo.active) {
-                            lizMap.controls.featureInfo.activate();
-                        }
-                    }
-                }
-            }
-        );
-
-        /**
-         * Polygon
-         * @type @new;OpenLayers.Control.DrawFeature
-         */
-        const queryPolygonLayerCtrl = new OpenLayers.Control.DrawFeature(
-            queryLayer,
-            OpenLayers.Handler.Polygon,
-            {
-                'featureAdded': onQueryFeatureAdded,
-                styleMap: drawStyleMap,
-                eventListeners: {
-                    // getFeatureInfo and polygon draw controls are mutually exclusive
-                    'activate': function() {
-                        if ('featureInfo' in lizMap.controls && lizMap.controls.featureInfo && lizMap.controls.featureInfo.active) {
-                            lizMap.controls.featureInfo.deactivate();
-                        }
-                    },
-                    'deactivate': function() {
-                        if ('featureInfo' in lizMap.controls && lizMap.controls.featureInfo && !lizMap.controls.featureInfo.active) {
-                            lizMap.controls.featureInfo.activate();
-                        }
-                    }
-                }
-            }
-        );
-
-        /**
-         * Freehand
-         * @type @new;OpenLayers.Control.DrawFeature
-         */
-        const queryFreehandLayerCtrl = new OpenLayers.Control.DrawFeature(queryLayer,
-            OpenLayers.Handler.Polygon, {
-                'featureAdded': onQueryFeatureAdded, styleMap: drawStyleMap,
-                handlerOptions: {freehand: true}
-            }
-        );
-
-        // TODO : keep reference to controls in this class
-        mainLizmap.lizmap3.map.addControls([queryPointLayerCtrl, queryLineLayerCtrl, queryPolygonLayerCtrl, queryBoxLayerCtrl, queryCircleLayerCtrl, queryFreehandLayerCtrl]);
-
-        mainLizmap.lizmap3.controls['selectiontool']['queryPointLayerCtrl'] = queryPointLayerCtrl;
-        mainLizmap.lizmap3.controls['selectiontool']['queryLineLayerCtrl'] = queryLineLayerCtrl;
-        mainLizmap.lizmap3.controls['selectiontool']['queryPolygonLayerCtrl'] = queryPolygonLayerCtrl;
-        mainLizmap.lizmap3.controls['selectiontool']['queryBoxLayerCtrl'] = queryBoxLayerCtrl;
-        mainLizmap.lizmap3.controls['selectiontool']['queryCircleLayerCtrl'] = queryCircleLayerCtrl;
-        mainLizmap.lizmap3.controls['selectiontool']['queryFreehandLayerCtrl'] = queryFreehandLayerCtrl;
-
-        mainLizmap.lizmap3.events.on({
-            'minidockopened': (mdoEvt) => {
-                if (mdoEvt.id == 'selectiontool') {
-                    this.toolSelected = 'deactivate';
-                    mainLizmap.lizmap3.layers['selectionQueryLayer'].destroyFeatures();
-                    mainLizmap.lizmap3.layers['selectionBufferLayer'].destroyFeatures();
+        // Listen to digitizing tool to query a selection when a feature is drawn
+        mainEventDispatcher.addListener(
+            () => {
+                for (const featureType of this.allFeatureTypeSelected) {
+                    mainLizmap.lizmap3.selectLayerFeaturesFromSelectionFeature(featureType, mainLizmap.digitizing.featureDrawn, this._geomOperator);
                 }
             },
-            'minidockclosed': (mdcEvt) => {
-                if (mdcEvt.id == 'selectiontool') {
-                    this.toolSelected = 'deactivate';
-                    mainLizmap.lizmap3.layers['selectionQueryLayer'].destroyFeatures();
-                    mainLizmap.lizmap3.layers['selectionBufferLayer'].destroyFeatures();
-                }
-            },
-            'layerSelectionChanged': () => {
-                mainEventDispatcher.dispatch('selectionTool.selectionChanged');
-            },
-            'layerFilteredFeaturesChanged': (lffcEvt) => {
-                if ($('#mapmenu li.selectiontool').hasClass('active') &&
-                    this.allFeatureTypeSelected.includes(lffcEvt.featureType)) {
-                    mainEventDispatcher.dispatch('selectionTool.filteredFeaturesChanged');
-                }
-            }
-        });
+            ['digitizing.featureDrawn']
+        );
     }
 
     get layers() {
@@ -392,15 +165,6 @@ export default class SelectionTool {
         if (btnSelectionTool.parentElement.classList.contains('active')) {
             btnSelectionTool.click();
         }
-    }
-
-    toggleVisibility() {
-        this._toogleSelectionLayerVisibility = !this._toogleSelectionLayerVisibility;
-
-        mainLizmap.lizmap3.layers['selectionQueryLayer'].setVisibility(this._toogleSelectionLayerVisibility);
-        mainLizmap.lizmap3.layers['selectionBufferLayer'].setVisibility(this._toogleSelectionLayerVisibility);
-
-        mainEventDispatcher.dispatch('selectionTool.toogleSelectionLayerVisibility');
     }
 
     unselect() {
